refactor(risk): relocate KillSwitchStore driven port to domain/ports (D1)
